+++ src/main/java/com/petillo/bankingapp/services/ClientService.java
@@ -1,6 +1,8 @@
 package com.petillo.bankingapp.services;
 
+import com.petillo.bankingapp.models.Account;
 import com.petillo.bankingapp.models.Client;
+import com.petillo.bankingapp.repositories.AccountRepo;
 import com.petillo.bankingapp.repositories.ClientRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -13,15 +15,17 @@ import java.util.List;
 public class ClientService {
 
     private final ClientRepo clientRepo;
+    private final AccountRepo accountRepo;
 
     private final String itemNotFound = "Client not found.";
 
     @Autowired
-    public ClientService(ClientRepo clientRepo){
+    public ClientService(ClientRepo clientRepo, AccountRepo accountRepo){
         this.clientRepo = clientRepo;
+        this.accountRepo = accountRepo;
     }
 
-    public Client getClientById(String id){
+    public Client getClientById(int id) throws ResponseStatusException{
         return clientRepo.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, itemNotFound));
     }
 
@@ -33,12 +37,21 @@ public class ClientService {
         return clientRepo.save(newClient);
     }
 
-    public Client updateClient(Client updatedClient){
-        return clientRepo.save(updatedClient);
+    public Client updateClient(int id, Client updatedClient) throws ResponseStatusException{
+        return clientRepo.findById(id)
+                .map(client -> {
+                    client.setFName(updatedClient.getFName());
+                    client.setLName(updatedClient.getLName());
+                    return clientRepo.save(client);
+                })
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, itemNotFound));
     }
 
-    //Add a check so that clients who have accts cannot be deleted
-    public void deleteClientById(String id){
+    public void deleteClientById(int id) throws ResponseStatusException{
+        List<Account> clientAccountList = accountRepo.findAllByClientId(id);
+        if (!clientAccountList.isEmpty()){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Client has active accounts. Deletion not allowed.");
+        }
         clientRepo.deleteById(id);
     }
